Minor name changes and reorganization for clarity
